Add files via upload
The gym chain Model Fitness is using data analysis to come up with a customer retention strategy, focusing on addressing customer churn. In the case of gym, a customer who is not visting gym during the whole month, generally suggests churn. Model Fitness has a digital client database which could be used to analyze churn cause.
In this project we need:

To redict the probability of churn for each customer in the upcoming month

Describe a typical user by selecting the most outstanding groups and narrating their main features

To analyze the factors impacting churn the most

Provide basic conclusions and develop recommendations on how to improve customer service:

o Identify target groups

o Suggest measures to cut churn

o Describe any other patterns with respect to interaction with customers